+++ src/py/hostname.rs
@@ -31,6 +31,8 @@ pub struct HostnameAnalysis {
     #[pyo3(get)]
     pub bidi_conflict: bool,
     #[pyo3(get)]
+    pub bidi_control: bool,
+    #[pyo3(get)]
     pub cross_label_script: bool,
     #[pyo3(get)]
     pub label_scripts: Vec<Vec<String>>,
@@ -50,6 +52,7 @@ impl From<crate::api::HostnameAnalysis> for HostnameAnalysis {
             mixed_script: a.mixed_script,
             has_confusables: a.has_confusables,
             bidi_conflict: a.bidi_conflict,
+            bidi_control: a.bidi_control,
             cross_label_script: a.cross_label_script,
             label_scripts: a.label_scripts,
             whole_script_confusable: a.whole_script_confusable,